all behaviours completed. success, failed by balance, failed by stock

+++ src/subscribe.rs
@@ -39,7 +39,7 @@ fn next() -> Option<PaymentTask> {
             )
             AND (
                 processing = false OR
-                updated_at < CURRENT_TIMESTAMP - INTERVAL '5 SEC'
+                updated_at < CURRENT_TIMESTAMP - INTERVAL '1 SEC'
             )
             ORDER BY next_try_at ASC, id ASC
             FOR UPDATE SKIP LOCKED
@@ -55,7 +55,7 @@ fn next() -> Option<PaymentTask> {
     result
 }
 
-fn success(payment_task: PaymentTask, payment: Payment, new_balance: i32) -> () {
+fn success(payment_task: PaymentTask, payment: Payment, product: Product, new_balance: i32) -> () {
 
     let desc: String = db::connection_desc();
     let mut pg = match Client::connect(&desc, NoTls) {
@@ -74,6 +74,16 @@ fn success(payment_task: PaymentTask, payment: Payment, new_balance: i32) -> () 
         panic!("{:?}", err_update);
     }
 
+    println!("Successful payment. New Balance: {:?}", new_balance);
+
+    // update product with new stock
+    let new_stock = (product.stock - 1);
+    println!("New Stock: {:?}", new_stock);
+    let mut q = "update products set stock = $1 where id = $2";
+    if let Err(err_update) = pg.execute(q, &[&new_stock, &product.id]) {
+        panic!("{:?}", err_update);
+    }
+
     // remove message from queue
     let query = "delete from payment_tasks where id = $1";
     if let Err(err_delete) = pg.execute(query, &[&payment_task.id]) {
@@ -81,16 +91,50 @@ fn success(payment_task: PaymentTask, payment: Payment, new_balance: i32) -> () 
     }
 }
 
-fn failed (payment_task: PaymentTask, user: User, payment: Payment) -> () {
+fn failed_by_balance (payment_task: PaymentTask, user: User, payment: Payment, product: Product) -> () {
+
+    let balance: i32 = user.balance;
+    let price: i32 = product.price;
+    let msg = format!("Unable to pay because price: {} is greater than balance {}", price, balance);
+
     let desc: String = db::connection_desc();
     let mut pg = match Client::connect(&desc, NoTls) {
         Ok(pg) => pg,
         Err(err_desc) => panic!("{:?}", err_desc)
     };
+
+    let mut q = "update payment_tasks set error = $1 where id = $2";
+    if let Err(err_update) = pg.execute(q, &[&msg, &payment_task.id]) {
+        panic!("{:?}", err_update);
+    }
+
+    let mut q = "update payments set status = 'rejected' where id = $1 ";
+    if let Err(err_update) = pg.execute(q, &[&payment.id]) {
+        panic!("{:?}", err_update);
+    }
+}
+
+fn failed_by_stock (payment_task: PaymentTask, user: User, payment: Payment) -> () {
+    let msg = "Unable to pay because there's no stock";
+    let desc: String = db::connection_desc();
+    let mut pg = match Client::connect(&desc, NoTls) {
+        Ok(pg) => pg,
+        Err(err_desc) => panic!("{:?}", err_desc)
+    };
+
+    let mut q = "update payment_tasks set error = $1 where id = $2";
+    if let Err(err_update) = pg.execute(q, &[&msg, &payment_task.id]) {
+        panic!("{:?}", err_update);
+    }
+
+    let mut q = "update payments set status = 'rejected' where id = $1 ";
+    if let Err(err_update) = pg.execute(q, &[&payment.id]) {
+        panic!("{:?}", err_update);
+    }
 }
 
 fn perform(payment_task: PaymentTask) -> () {
-    println!("{:?}", payment_task);
+    println!("Processing Task: {:?}", payment_task);
 
     let desc: String = db::connection_desc();
     let mut pg = match Client::connect(&desc, NoTls) {
@@ -103,30 +147,36 @@ fn perform(payment_task: PaymentTask) -> () {
         Ok(payment) => Payment::from(payment),
         Err(desc) => panic!("No payment was found")
     };
-    println!("{:?}", payment);
+    println!("Payment: {:?}", payment);
 
     let mut q = "select * from users where id = $1";
     let user = match pg.query_one(q, &[&payment.user_id]) {
         Ok(user) => User::from(user),
         Err(desc) => panic!("no user found")
     };
-    println!("{:?}", user);
+    println!("for User: {:?}", user);
 
     let mut q = "select * from products where id = $1";
     let product = match pg.query_one(q, &[&payment.product_id]) {
         Ok(product) => Product::from(product),
         Err(desc) => panic!("no user found")
     };
-    println!("{:?}", user);
+    println!("{:?}", product);
 
     let balance: i32 = user.balance;
     let price: i32 = product.price;
-    let new_balance: i32 = balance - price;
+    let stock: i32 = product.stock;
 
-    if new_balance < 0 {
-        failed(payment_task, user, payment)
+    let new_balance: i32 = balance - price;
+    let new_stock: i32 = stock - 1;
+
+    if (new_balance < 0) {
+        failed_by_balance(payment_task, user, payment, product)
+    } else if (new_stock < 0) {
+
+        failed_by_stock(payment_task, user, payment)
     } else {
-        success(payment_task, payment, new_balance)
+        success(payment_task, payment, product, new_balance)
     }
 
 }
